requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
